Organize test files into folders mirroring the source layout
Move the flat test files into Services/{Install,Feeds,News,
Infrastructure}, ViewModels, Models, and Patching folders matching the
structure of the code under test, with namespaces to match. TestHelpers
and AppManifestTest stay at the root (cross-cutting), HeadlessTests was
already grouped.
