+++ tests/Ksp2Redux.Tools.Launcher.Tests/Services/Feeds/ManifestReleasesFeedUpdateManifestTest.cs
@@ -5,7 +5,7 @@ using Ksp2Redux.Tools.Launcher.Services.Infrastructure;
 using Moq;
 using Testably.Abstractions.Testing;
 
-namespace Ksp2Redux.Tools.Launcher.Tests;
+namespace Ksp2Redux.Tools.Launcher.Tests.Services.Feeds;
 
 public class ManifestReleasesFeedUpdateManifestTest
 {
