+++ tests/Ksp2Redux.Tools.Launcher.Tests/Services/Install/Ksp2InstallServiceTest.cs
@@ -4,7 +4,7 @@ using Ksp2Redux.Tools.Launcher.Services.Install;
 using Ksp2Redux.Tools.Launcher.Services.Infrastructure;
 using Moq;
 
-namespace Ksp2Redux.Tools.Launcher.Tests;
+namespace Ksp2Redux.Tools.Launcher.Tests.Services.Install;
 
 public class Ksp2InstallServiceTest
 {
