+++ tests/Ksp2Redux.Tools.Launcher.Tests/Services/News/NewsServiceTest.cs
@@ -3,7 +3,7 @@ using Ksp2Redux.Tools.Launcher.Services.News;
 using Ksp2Redux.Tools.Launcher.Services.Infrastructure;
 using Moq;
 
-namespace Ksp2Redux.Tools.Launcher.Tests;
+namespace Ksp2Redux.Tools.Launcher.Tests.Services.News;
 
 public class NewsServiceTest
 {
